Checking if objects are what they supposed to be.

+++ module/Application/src/Application/Controller/RepoController.php
@@ -34,6 +34,13 @@ class RepoController extends AbstractActionController
             $repository = $sm->get('EdpGithub\Client')->api('repos')->show($owner, $repo);
             $repository = json_decode($repository);
 
+            if(!($repository instanceOf \stdClass)) {
+                throw new Exception\RuntimeException(
+                    'Not able to fetch the repository from github due to an unknown error.',
+                    500
+                );
+            }
+
             if(!$repository->fork && $repository->permissions->push) {
                 if($this->isModule($repository)) {
                     $service = $this->getModuleService();
@@ -66,7 +73,7 @@ class RepoController extends AbstractActionController
      * @throws Exception\UnexpectedValueException
      * @return
      **/
-   public function removeAction()
+    public function removeAction()
     {
         $request = $this->getRequest();
         if($request->isPost()) {
@@ -77,11 +84,25 @@ class RepoController extends AbstractActionController
             $repository = $sm->get('EdpGithub\Client')->api('repos')->show($owner, $repo);
             $repository = json_decode($repository);
 
+            if(!$repository instanceOf \stdClass) {
+                throw new Exception\RuntimeException(
+                    'Not able to fetch the repository from github due to an unknown error.',
+                    500
+                );
+            }
+
             if(!$repository->fork && $repository->permissions->push) {
                 $mapper = $sm->get('application_module_mapper');
                 $module = $mapper->findByUrl($repository->html_url);
-                $module = $mapper->delete($module);
-                $this->flashMessenger()->addMessage($repository->name .' has been removed from ZF Modules');
+                if($module instanceOf \Application\Entity\Module) {
+                    $module = $mapper->delete($module);
+                    $this->flashMessenger()->addMessage($repository->name .' has been removed from ZF Modules');
+                } else {
+                    throw new Exception\UnexpectedValueException(
+                        $repository->name . ' was not found',
+                        403
+                    );
+                }
             }else {
                 throw new Exception\UnexpectedValueException(
                     'You have no permission to add this module. The reason might be that you are' .
@@ -107,13 +128,13 @@ class RepoController extends AbstractActionController
     {
         $sm = $this->getServiceLocator();
         $client = $sm->get('EdpGithub\Client');
-        $em = $client->getHttpClient()->getEventManager();
-        $errorListener = $sm->get('EdpGithub\Listener\Error');
-        $em->detachAggregate($errorListener);
-        $module = $client->api('repos')->content($repo->owner->login, $repo->name, 'Module.php');
-        $response = $client->getHttpClient()->getResponse();
-        $em->attachAggregate($errorListener);
-        if(!$response->isSuccess()){
+        try{
+            $module = $client->api('repos')->content('someone', 'someRepo', 'Module.php');
+        } catch(\Exception $e) {
+            return false;
+        }
+
+        if(!json_decode($module) instanceOf \stdClass) {
             return false;
         }
